pyproject.toml
[tool.poetry]
name = "mechanisms-of-action-prediction"
version = "0.1.0"
description = ""
authors = ["navitacion"]

[tool.poetry.dependencies]
python = "^3.8"
pytorch_lightning = "^0.9.0"
hydra-core = "^1.0.3"
comet_ml = "^3.2.2"
pandas = "^1.1.2"
scipy = "^1.5.2"
matplotlib = "^3.3.2"
seaborn = "^0.11.0"
scikit-learn = "^0.23.2"
jupyter = "^1.0.0"
jupyterlab = "^2.2.8"
opencv-python = "^4.4.0"
category_encoders = "^2.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
